Further generalize base Ridwell entity
